LightboxFooter: Make layout more consistent with the header.
Now that we've trimmed away a few unnecessary properties in
`styles.wrapper` in LightboxHeader, it's easier to see how it
differs from the same set of styles in LightboxFooter. Staring at
those differences, some of them don't make a lot of sense, and it
seems like an opportunity to give the layout a more consistent feel.

Make the header and the footer more consistent by

- making the three-dots icon and the the "X" icon the same size

- using the same horizontal padding, so the three-dots icon aligns
  with the "X" icon in the header

- using the same vertical padding, instead of the arbitrary
  `height: 44`, for the vertical layout