fetchActions: Fix recent regression in post-initial-fetch navigation.
In e18124898, we forgot to change this instance of 'main' to
'main-tabs'. The symptom was that you could navigate somewhere in
the app (e.g., to the message list) while the "Connecting..." banner
was showing, then be interrupted when that banner disappeared by
having the nav reset back to the home tab -- which was annoying.