Revert "readme: Fix some broken links."
This reverts commit ff0b462a5.

As Greg says [1],

"""
Hmm - there's actually a reason those links were to that text and
not directly to the chat stream. That section of text is where we
link to the document with norms and conventions for the
chat.zulip.org community:
https://zulip.readthedocs.io/en/latest/contributing/chat-zulip-org.html

There's important context there that I'd like people to see when
they're first joining.
"""

[1] https://github.com/zulip/zulip-mobile/pull/4456#issuecomment-771192432