fetchActions: Remove a now-unnecessary navigation.
This is fine because we've just changed all the places where we were
navigating to 'loading', to instead navigate to 'main-tabs'.

This removes a `NavigationService` callsite that's been throwing
quite a lot; that's #4453. We still don't know why it's throwing
those errors, and we haven't managed to reproduce them yet. But it
means we should continue to prioritize removing `NavigationService`.

See discussion at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.23M4453.20.22Tried.20to.20use.20NavigationService.20before.20appContaine.2E.2E.2E/near/1111875.

Fixes: #4453