MessageList [nfc]: Require `messages` and `initialScrollMessageId`.
This change in `MessageList`'s interface might be seen as stretching
a bit far to accommodate the search screen's peculiar choices. This
is partly fair, but also, the choice to keep reusing a widget in two
important places means we should concede, to some extent, that one
use isn't peculiar just because it's different from the other.