PasswordAuthScreen: Use new `realm` route param.
And stop using `getPartialAuth`. Instead, fill out an `Auth` object
with data that can be known without grabbing anything from Redux.
Appropriately (since the point of this screen is to get an API key),
the place where we can't use anything better than the empty string
for `apiKey` is also a place where a proper API key isn't necessary.

Fixes: #4415