accountsSelectors [nfc]: Remove `getPartialAuth`.
We've just removed the only callers, which were all in the "for use
in authentication flows" usage category. No callers were in the
other category: "or other places that operate on an active account
which may not be logged in". This latter category seems vaguely
useful for when we make improvements to multi-account support, but
nothing specific comes to mind, and it seems best to let our future
selves decide what we'll need, when we know our requirements a bit
better.