deps: Upgrade @sentry/react-native to 2.1.1, the latest.
One thing we'd like to have is the merged PR
getsentry/sentry-javascript#3108, a type adjustment that lets us use
`undefined` as a value in `setTag` and `setTags`. The author of that
PR says that passing `undefined` for a tag's value is an unofficial
way to unset a tag from the scope. If that works, we'd like to do it
until getsentry/sentry-javascript#2218 is solved more officially.

I see just one declared breaking change at 2.0.0 [1]:

"""
- build(android): Changes android package name from
  `io.sentry.RNSentryPackage` to `io.sentry.react.RNSentryPackage`
  (Breaking). #1131
"""

But autolinking handles the name change just fine, so there's
nothing special we have to do about that.

We don't have a complete libdef yet, but we seem to have good
coverage for our usage in what we've translated from the TypeScript
manually so far. So, double-check all of that against the new
version and fix a few things that were slightly wrong or out-of-date
(including by taking the relevant changes from
getsentry/sentry-javascript#3108).

[1] https://github.com/getsentry/sentry-react-native/blob/master/CHANGELOG.md#200