ComposeBox: Reset the topic input more sensibly when finished editing.
`lastMessageTopic` seems to be our go-to for a sensible default
value for the topic input. It'll be the empty string (and thus NFC)
in all cases except when the last message in the narrow exists and
is a stream message. In that case it seems helpful to suggest an
existing topic, instead of no topic.

In practice, while I have in the past noticed a bug (without finding
a consistent reproduction recipe) where the topic input is empty
when I don't expect it to be, I think that's the exception rather
than the rule. Still, I suspect this is due to the same (or a
similar) leaning tower of hacks as I described in steps 1-5 in the
previous commit, so not good to depend on.