exampleData: Have `pmMessage` set `sender` based on a passed `sender_id`.
The immediate motivation is to resolve a complaint from Flow about a
situation that "may lead to an exponentially large number of cases
to reason about", when we upgrade to Flow v0.128.

But also, I think `sender_id` makes more sense as an alias for
passing a whole unique `sender` with a specific ID, than an alias
for passing `otherUser` with a changed ID field. I found nothing to
contradict this in zulip/zulip-mobile@f59cd486e1, when `sender_id`
was added.