getMessages types: Make `messages` in `migrateMessages` a read-only array.
There's no reason we'd want to mutate this array.

Given that, and the fact that we're about to make `ServerMessage` a
union of `ServerMessageOf<PmMessage>` and
`ServerMessageOf<StreamMessage>`, we might as well be courteous to
`migrateMessages`' callers by allowing them to pass an array of just
one of those types. In other words, we might as well allow this
argument, which is an array, to be treated covariantly in the type
of its elements. This is only possible when we the array is
read-only.

See discussion at
  https://github.com/zulip/zulip-mobile/pull/4222#discussion_r468263756,
and in particular the example in "What about other types" at the
bottom of this article:
  https://www.stephanboyer.com/post/132/what-are-covariance-and-contravariance.