webview: Skip showing user-agent in red error banner.
This banner is generally shown only in development -- in release
builds, we just show a terser, friendlier, generic error message.
In a development context there are plenty of other ways to confirm
the browser version.  And on the other hand the user-agent string
tends to be quite long, adding several lines to the height of the
banner, which makes the screen pretty crowded and could make it
harder to see what's going on for debugging.

So, leave it out from the banner, but just include it in the error
report sent to Sentry.

Also add some comments, because this code isn't self-explanatory
about that difference in behavior.