sharing: Make the stream/topic input list scrollable.
Previously the scroll event was getting registered by the ScrollView
enclosing these input list, hence users were not able to scroll
through the input list provided in the sharing screen.

Upstream docs:
  https://reactnative.dev/docs/scrollview#nestedscrollenabled-android
say this defaults to true, which should mean this is redundant.
But empirically, we seem to need it.  Discussion at:
  https://github.com/zulip/zulip-mobile/pull/4514#discussion_r620895824