flow: Upgrade to v0.126.
As noted in a recent commit where we upgraded to Flow v0.125, we
don't usually bump the Flow version outside of a React Native
upgrade commit.

But I found that we could get v0.126 without any added Flow errors
in React Native code -- which is even better than the situation with
Flow v0.125, where we saw one Flow error in one file.

Being on v0.126 is nice because we'll finally be able to drop our
comments like "this should really be exact -- see note in
jsonable.js" on many objects with indexer properties, and make those
object types exact.

- That, in turn, will help us address #3452 more productively
  ("Convert all object types to exact, or explicitly-inexact").

- Fixing #3452 is really a prerequisite for taking
  facebook/react-native@050a7dd01 (switching on `exact_by_default`),
  which is on the path to the RN v0.64 upgrade.

[1] https://flow.org/en/docs/config/options/#toc-exact-by-default-boolean