logo [nfc]: Simplify a bit by sharing some variables with callee.
This function make_one_android_icon takes 6 parameters, which is
well past the point where they should be named instead of positional.

The Bash language doesn't have a great way to pass named arguments
to a function... but this is about the closest we can get.  The major
downside is that it's not visible at the call site that this is what's
happening.  For this local helper, where the only call sites are just
below the function itself and indeed fit on a screen with it, that's
not catastrophic; and it seems worth the improvement in readability of
the parts that do vary from call site to call site.