android: Fix issue where another instance of app would start on sharing.
This would occur whenever zulip was started by another Activity using
startActivityForResult. Ideally with launchMode="singleTask" zulip
should always launch in a seperate task, this however is not the case,
when it is started via startActivityForResult by another app.
With startActivityForResult another instance of zulip would be
initialized inside the calling app, regardless of the fact that zulip
was in background or not.

This caused problems especially for share to zulip intents and this
commit addresses those cases. Now SEND intent is not captured by
MainActivity, a seperate native activity is created to capture them
and it redirects the intents unchanged to MainActivity by calling
startActivity, hence always ensuring that zulip is launched as a
seperate task. (If zulip was already running and in background, it will
be used instead of creating a new task.)

Details on how I came to this conclusion:

- https://github.com/zulip/zulip-mobile/pull/4514#issuecomment-798069185