user: Expand checks in getHaveServerData.
The one check we had would be perfectly sufficient in a world where
we used a reliable form of storage for keeping data from the server
on the device between runs of the app.  It does not suffice given
our model with redux-persist.

This doesn't fix the problem -- that will take something more
fundamental -- but it hopefully mitigates the worst consequences.

In particular, empirically we already reasonably gracefully handle
in the unreads view and the PM-conversations view a situation where
either `state.unreads` or `state.pmConversations` refers to streams,
subscriptions, or users that we don't have data about:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/Android.3A.20White.20screen/near/1151982

Fixes: #4587