ComposeMenu: Don't immediately send attachment on upload.
On pressing attach file from compose menu, attachment was directly
dispatched to outbox, they should go through compose box so that
a user may add more text if desired.

This commit creates a new function `insertAttachment` that is
utilized in `ComposeMenu`. It performs the task of uploading the
attachment as well as updating the `ComposeBox` message accordingly.

Fixes part of #4540 because behaviour of image picker option is
unchanged.

Fixes-part-of: #4540