WebLink: Require an absolute, already validated URL.
A crash on PasswordAuthScreen was recently introduced, in #4491:
`getCurrentRealm` requires a current realm from Redux, and we
stopped storing that during this in-between state when the user
hasn't actually logged in yet.

The crash wouldn't have been introduced if this component weren't
sensitive to there being a current realm in Redux or not. So, rather
than choosing a solution that keeps some of that sensitivity, just
require callers to pass the full, absolute, validated URL.

Fixes: #4605