events: Fix bug of stale state across a batch of events.
This is the bug we discovered in the previous commit as a result
of juxtaposing this code.  Instead of getting the state up front
for a given batch of events once and for all, we should convert
each event to an action using the new state after dispatching any
previous actions.

In order to get the flexibility to do this correctly, we fuse the
map/filter stuff into the plain old explicit loop.