AppNavigator: Use `withHaveServerDataGate` for most screen components.
The user could have any of these screens mounted when they receive a
force-logout. If that happens, we want them all to handle it
gracefully.

Putting the calls here, instead of alongside each component's
definition, will make it easier to scan through and spot any that
should use the HOC and don't.