queueMarkAsRead: Resend messages on failure.
Previously, if a `messagesFlags` message failed to send, we would ignore
the error, and not retry the request until the user happened to do
something that would trigger marking those messages as read again.

This changes the code to try resending the `messagesFlags` message for
every ID we try to mark read, until the server explicitly tells us that
it has succeeded.

If we kept around this queue forever, we would want to do something like
detect if an ID has been retried more than some number of times, or for
some amount of time, and give up, but given that this is currently an
in-memory queue, it seems fine to keep trying until the app happens to
restart.

I tested this by going into a topic with many unreads, enabling airplane
mode, scrolling down, then disabling airplane mode. Previously, you'd
need to scroll all the way back up to the first message that failed,
marking each message as read again. Now, when you scroll back up, as
soon as you hit a single message that's unread, it marks them all as
read.