UploadedAvatarURL [nfc]: Remove global flag in regex.
Even in cases where there are multiple matches (and such cases seem
unlikely), we've already opted out of any behavior changes made
possible by the `g` flag here: we only call this `RegExp` object
once before discarding it, and it'll return the first match from
that call, which is the same as it'd do without the `g`. We don't
then go and ask it to find subsequent matches. See Greg's demo and
discussion at
  https://github.com/zulip/zulip-mobile/pull/4625#discussion_r610224024.

Soon, we'd like to remove the `if (match !== null)` check and give
`String.prototype.replace` the responsibility of executing the
`RegExp` object. Since that method does some fancy stuff with a `g`
flag, it's best to cut it out now, before we make that change.

In particular, that fancy stuff is described by MDN [1]:

- "To perform a global search and replace, include the `g` switch in
  the regular expression."

- "Note that the function will be invoked multiple times for each
  full match to be replaced if the regular expression in the first
  parameter is global."

It'd replace multiple matches in the string, which is different from
what happens now, and not what we want.

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/replace