SelectableOptionRow [nfc]: Add, to replace `LanguagePickerItem`.
Created by copy-pasting `LanguagePickerItem`.

In the next several commits, we'll make this into a more
general-purpose component that may be useful to represent a checkbox
or a radio button.

Then we'll switch `LanguagePickerItem`'s caller to use this
component instead, and remove `LanguagePickerItem`. We'll also use
it for the items on `UserStatusScreen`, which now incorrectly show
right-facing arrows, as though they'll navigate somewhere.

Future uses of this component (e.g., we may want it for #4009) can
help refine its interface. For now, just let that process start --
and make it easier to provide a consistent look-and-feel -- by
making a reusable component.