emoji: Fix :a:, :b:, :o:, :p: to have emoji-style presentation.
This causes them to render (at least on Android) with bigger, colored
glyphs that look a lot like the glyphs we use in the webapp.

Without this, they render with text-style glyphs: transparent letter
on white background, which also means that in a light theme they're
completely invisible.

These are the only emoji we have in this range of letter characters.
There are some others like :ok: that are from neighboring characters
representing multiple letters, but those all seem to be doing fine
already.

Fixes: #3395