ZulipAsyncStorage types: Better align .setItem/.getItem with AsyncStorage.
Now, the type for the `callback` param matches what's in the libdef
for `AsyncStorage`. I think the question mark in the `?(...) => Foo`
syntax isn't what we'd normally choose, but `ZulipAsyncStorage` is a
wrapper around `AsyncStorage` so it might as well match this detail.

Also the callback in `getItem` wasn't being considered optional at
all, and now it is, which it looks like our implementation expects.