flipper ios: Pin to some specific versions, to fix build with Xcode 12.5.
Following the outlined solution for projects using RN v0.63, in
facebook/react-native#31480.

The "properly clean your caches" step didn't seem to be necessary
for me, but possibly that's different between environments.

Done without looking for breaking changes, beyond confirming that
the build succeeds on Xcode 12.5. There could be a lot of breaking
changes between Flipper 0.54.0 and 0.75.1 -- but we don't use
Flipper yet (that's #4255) and we don't have anything like a Flipper
config file for iOS.