getInitialRouteInfo: Give 'account-pick' when just one, logged-out account.
Instead of 'realm-input'.

The entire `!hasAuth` condition may become impossible, in a world
where we don't keep around accounts that aren't logged in [1]. So,
even if it's not totally clear that the new behavior from this
commit is an improvement, at least we know that it might not be
permanent. (I do think it's an improvement, for a reason I describe
in a comment.)

This removes the second of the two cases of #4604; see the comment
in the removed `.length === 1` case describing the
`RealmInputScreen` flash as the expected behavior.

[1] https://chat.zulip.org/#narrow/stream/48-mobile/topic/logged-out.20state/near/1174840

Fixes: #4604