MainTabsScreen [nfc]: Make it clear that `isPad` is iOS-only.
Although it doesn't give a reason why, the `Platform` doc [1] is
clear that `isPad` is iOS-only; we can expect it to be missing on
Android.

We'd love to know whether an Android user is on a tablet, but it
seems `Platform` doesn't yet give us a way to do that. Until it
does, or until we use something else that does, make it really clear
that we're not getting accurate info on Android.

[1] https://reactnative.dev/docs/platform#ispad-ios