types: Distinguish PmOutbox vs. StreamOutbox more fully
This expresses that the types of these properties `type`, `subject`,
and `display_recipient` vary together, and that they do so in the
same way as on `Message`.

That in turn means that when we have a condition like `type ===
'stream'`, Flow can infer that `display_recipient` will be a string
(the stream name) rather than an array, and so on.