redux-persist [nfc]: Remove a noisy log line in `autoRehydrate`.
This wasn't being logged anyway, since we haven't been passing
`config.log` with the `autoRehydrate` call so far.

But it's something that would get logged when things work as
expected, so it'd be really noisy. If someone needs to log stuff for
debugging, they'll want to find their own targeted logs to use
anyway.