deps: Run `yarn upgrade`.
Notably, this includes an update to Sentry's SDK for React Native
that came out after the last `yarn upgrade` commit, in #4789, was
authored.

I got an error --

```
[!] CocoaPods could not find compatible versions for pod "Sentry":
  In snapshot (Podfile.lock):
    Sentry (= 6.1.4)

  In Podfile:
    RNSentry (from `../node_modules/@sentry/react-native`) was resolved to 2.5.0, which depends on
      Sentry (= 7.0.0)

You have either:
 * out-of-date source repos which you can update with `pod repo update` or with `pod install --repo-update`.
 * changed the constraints of dependency `Sentry` inside your development pod `RNSentry`.
   You should run `pod update Sentry` to apply changes you've made.
```

-- so I ran `pod update Sentry`, and it went through. After that, I
ran `yarn` (with `pod install` getting run in the postinstall
script), and no further changes were made.

We should aim to have this out in a beta for a little while, to
shake out any bugs. In addition to this Sentry upgrade, quite a lot
of other stuff seems to have been touched.