RawLabel: Fix crash when text is empty string.
This was causing a crash when trying to autocomplete a user group with a
blank name. The empty string was passed into RawLabel, which caused this
invariant to incorrectly trigger.

This appears to have been introduced in 199a19069.

Instead of using `!!` to check for undefined, we can do `!= null`, which
will trigger for undefined and null, but not an empty string.