ios build: Run "Update to recommended settings" for ZulipMobile project.
Done on Xcode "Version 12.5 (12E262)" in the GUI; screenshot at
  https://github.com/zulip/zulip-mobile/pull/4887#issuecomment-876554362.

Done to follow some of the template-app changes in
facebook/react-native@6685aba46, on the path to the RN v0.64
upgrade. In particular, this bullet point:

> - Update to recommended settings for Xcode 12

I didn't use a similar option that presented itself for the Pods
project.

(We don't make any changes corresponding to the "Add -e to the
bundle images script" bullet point; see
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/RN.20v0.2E64.20upgrade.3A.20Xcode.20config.20changes.3F/near/1209893.)