tools/test [nfc]: Inline random_platform.
No need to compute this in advance of where we decide we'll be
interested in the result.

This also means the funny `RANDOM` syntax goes right next to the
comment saying in English that yes, this chooses randomly.

While here, also fix the spacing in this array to match the other
similar lines around it.