deps: Follow RN in updating some dev-dependency ranges.
Done to follow the template-app change in
facebook/react-native@fffa4d1ae, on the path to the RN v0.64
upgrade.

I ignored the deps we don't use, and I ignored the ones for which we
already specify a higher lower bound.

Squinting at the `yarn.lock`, this doesn't actually change any of
the versions we actually end up using. It just bumps the explicit
lower bounds in the `package.json`.