lightbox [nfc]: Remove `fileCache` property from rnfb config.
This is not required as 'path' with `CacheDir` fulfills the same
purpose as `fileCache`; and passing `path` means that `fileCache`
has no effect [1].

The difference is 'path' gives specific path to the file while
`fileCache` gives a randomly generated path (including the
filename and its extension). See [2].

We would want to preserve the filename and its extension while
sharing from lightbox and hence `path` is used instead of
`fileCache`.

Originally reported at [3] by Greg.

reference:
[1] https://github.com/joltup/rn-fetch-blob/wiki/Classes#pathstring-1
[2] https://github.com/joltup/rn-fetch-blob#download-to-storage-directly
[3] https://chat.zulip.org/#narrow/stream/48-mobile/topic/Sharing.20from.20lightbox.20broken.20on.20Android/near/1226489