createPersistor: Stop ignoring state changes that happen during a write.
In an earlier commit in this series, we stopped blending snapshots
together when we initiate writes. A consequence to the approach we
took was to ignore all state changes that happen while a write is in
progress.

Now, we come back to fix that problem, but after reworking the logic
to use `AsyncStorage.multiSet`.

This finishes the approach outlined by Greg; see
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/data.20storage/near/1214393.