react-native-safe-area-context libdef: Use a built-in React type.
We already have a type wrapper for this function, from 8f569641f, so
we could just as well have suppressed type-checking for this whole
function here. But this is easy.

The Flow error that caused me to notice this appeared when I tried
to upgrade to Flow v0.137, for the RN v0.64 upgrade:

```
Error -------------------------------------------------------- flow-typed/react-native-safe-area-context_vx.x.x.js:60:23

Cannot use `React.ComponentType` as a type because it is an `any`-typed value. Type `React.ComponentType` properly, so
it is no longer `any`-typed, to use it as an annotation. [value-as-type]

   60|     WrappedComponent: React.ComponentType<T>,
                             ^^^^^^^^^^^^^^^^^^^
```