deps: Add caret on @react-native-async-storage/async-storage.
This takes us to v1.15.5, the current latest. No breaking changes
are announced in the changelog [1], but 1.13.3 does announce this:

> After upgrading to v1.13.3+, remember to:
>
> - remove Pods directory (ios/Pods)
> - clean build directory through Xcode (in menu: Product -> Clean
>   Build Folder)
> - install dependencies again (`pod install`)

All this didn't seem obviously necessary (except `pod install`,
which we always do), but I did it anyway.

v1.13.4 gets us @react-native-async-storage/async-storage@fb0ce97e3,
which we'll need for the RN v0.64 upgrade, #4426.

[1] https://github.com/react-native-async-storage/async-storage/releases

Fixes: #4890